+++ src/compiler/visitor.js
@@ -8,12 +8,13 @@ let visitor = {
     let step = function(node) {
       let type = node[0];
       if (type && fns[type]) {
-        return fns[type].apply(null, node);
+        return fns[type].apply(null, arguments);
       }
     };
     let walk = function(nodes=[]) {
+      let extraArgs = Array.prototype.slice.call(arguments, 1);
       nodes.forEach((n) => {
-        step(n);
+        step.apply(null, [n].concat(extraArgs));
       });
     };
